requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Graph and clustering detection
python-louvain>=0.16
networkx>=3.0

# Data processing
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
umap-learn>=0.5.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
requests>=2.28.0

# Optional: For downloading datasets
gitpython>=3.1.0
